requirements.txt
annotated-types==0.5.0
anyio==3.7.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.101.0
geographiclib==2.0
geopy==2.3.0
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
orjson==3.9.4
packaging==23.1
pluggy==1.2.0
psycopg2==2.9.7
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pytest==7.4.0
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.0
starlette==0.27.0
tomli==2.0.1
typing_extensions==4.7.1
ujson==5.8.0
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
websockets==11.0.3
